Restructures the UI modules
