+++ ui/src/app/modules/domain/thought.ts
