+++ ui/src/app/modules/domain/websocket-response.ts
