--- ui/src/app/modules/teams/components/feedback-form/feedback-form.component.html
@@ -1,37 +1,0 @@
-<!--
-  ~ Copyright (c) 2018 Ford Motor Company
-  ~ All rights reserved.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<div class="title">
-  Please share your thoughts and impressions with us. We're always improving this tool and your feedback is
-  appreciated.
-</div>
-
-<div class="rating">
-      <span *ngFor="let star of stars.reverse(); let i = index"
-            (click)="setRating(maxStars - i)">{{star}}</span>
-</div>
-<div>
-  <label for="comments">Comments</label>
-  <textarea id="comments" placeholder="Enter comments here" rows="5" [(ngModel)]="comments"></textarea>
-</div>
-<div>
-  <label for="email">Email</label>
-  <input id="email"
-         placeholder="Enter email if you would like us to follow up"
-         type="email"
-         [(ngModel)]="followupEmail"/>
-</div>
