--- ui/src/app/modules/teams/components/feedback-form/feedback-form.component.scss
@@ -1,77 +1,0 @@
-/*!
- * Copyright (c) 2018 Ford Motor Company
- * All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-@import 'color-vars';
-
-* {
-  box-sizing: border-box;
-}
-
-.title {
-  font-size: 1.25rem;
-  max-width: 55vw;
-}
-
-textarea {
-  border-color: $text-gray;
-  border-radius: 3px;
-  font-size: 1rem;
-  padding: 8px;
-  resize: none;
-  width: 100%;
-}
-
-label {
-  display: block;
-  font-weight: bold;
-  margin: 1rem 0;
-}
-
-input {
-  background-color: $white;
-  border: 1px solid $text-gray;
-  border-radius: 3px;
-  font-size: 1rem;
-  padding: 8px;
-  width: 100%;
-}
-
-.rating {
-  color: $button-blue;
-  direction: rtl;
-  font-size: 2rem;
-  padding: 25px 0 0;
-  unicode-bidi: bidi-override;
-  word-wrap: normal;
-
-  & > span {
-    cursor: pointer;
-    display: inline-block;
-    font-family: 'Material Icons', fantasy;
-    font-size: 3rem;
-    position: relative;
-    width: 1.1em;
-
-    // sass-lint:disable-all
-    // TODO: fix this plz
-
-    &:hover::before,
-    &:hover ~ span::before {
-      content: 'star';
-      position: absolute;
-    }
-  }
-}
