--- ui/src/app/modules/teams/components/feedback-modal/feedback-modal.component.html
@@ -1,30 +1,0 @@
-<!--
-  ~ Copyright (c) 2018 Ford Motor Company
-  ~ All rights reserved.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<rq-modal [isOpen]="isOpen" (close)="close()">
-  <div class="modal-body">
-    <rq-feedback-form #feedbackForm></rq-feedback-form>
-  </div>
-
-  <div class="modal-footer">
-    send feedback
-
-    <div class="action-section">
-      <i class="fas fa-times-circle close modal-button" (click)="close()"></i>
-      <i class="fas fa-check-circle submit modal-button" (click)="submit()"></i></div>
-  </div>
-</rq-modal>
