--- ui/src/app/modules/teams/components/feedback-modal/feedback-modal.component.ts
@@ -1,41 +1,0 @@
-/*
- * Copyright (c) 2018 Ford Motor Company
- * All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import {Component, Input, ViewChild} from '@angular/core';
-
-@Component({
-  selector: 'rq-feedback-modal',
-  templateUrl: './feedback-modal.component.html',
-  styleUrls: ['./feedback-modal.component.scss']
-})
-export class FeedbackModalComponent {
-
-  @Input() teamId: string;
-  isOpen = false;
-
-  @ViewChild('feedbackForm') feedbackForm;
-
-  close () {
-    this.isOpen = false;
-    this.feedbackForm.reset();
-  }
-
-  submit () {
-    this.feedbackForm.submit();
-    this.close();
-  }
-}
