--- ui/src/app/modules/teams/components/modal/modal.component.html
@@ -1,26 +1,0 @@
-<!--
-  ~ Copyright (c) 2018 Ford Motor Company
-  ~ All rights reserved.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<div class="modal-wrapper" [hidden]="!isOpen">
-  <div class="modal">
-    <ng-content select=".modal-header"></ng-content>
-    <ng-content select=".modal-body"></ng-content>
-    <ng-content select=".modal-footer"></ng-content>
-  </div>
-
-  <div class="modal-background" (click)="closeModal()"></div>
-</div>
