--- ui/src/app/modules/teams/components/modal/modal.component.scss
@@ -1,83 +1,0 @@
-/*!
- * Copyright (c) 2018 Ford Motor Company
- * All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-@import 'color-vars';
-
-.modal-wrapper {
-  align-items: center;
-  bottom: 0;
-  display: flex;
-  justify-content: center;
-  left: 0;
-  padding: 20px;
-  position: fixed;
-  right: 0;
-  top: 0;
-  z-index: 99;
-
-  &[hidden] {
-    display: none;
-  }
-
-  .modal {
-    background-color: $white;
-    border-radius: 6px;
-    box-shadow: 0 19px 38px opacity($black, .3), 0 15px 12px opacity($black, .22);
-    margin-top: -10vh;
-    width: 70%;
-    word-wrap: break-word;
-    z-index: 100;
-
-    &::ng-deep {
-      .modal-body {
-        font-size: 1rem;
-        padding: 1.4rem;
-      }
-
-      .modal-header {
-        background-color: $button-blue;
-        border-radius: 6px 6px 0 0;
-        padding: 1.4rem;
-
-        &:empty {
-          display: none;
-        }
-      }
-
-      .modal-footer {
-        background-color: $button-blue;
-        border-radius: 0 0 6px 6px;
-        color: $white;
-        padding: 1.4rem;
-        text-transform: uppercase;
-
-        &:empty {
-          display: none;
-        }
-      }
-    }
-  }
-
-  .modal-background {
-    background-color: opacity($modal-blur, .8);
-    bottom: 0;
-    left: 0;
-    position: fixed;
-    right: 0;
-    top: 0;
-    z-index: 95;
-  }
-}
