--- ui/src/app/modules/teams/components/modal/modal.component.spec.ts
@@ -1,49 +1,0 @@
-/*
- * Copyright (c) 2018 Ford Motor Company
- * All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import {ModalComponent} from './modal.component';
-
-describe('ModalComponent', () => {
-  let component: ModalComponent;
-
-  beforeEach(() => {
-    component = new ModalComponent();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-
-  describe('closeModal', () => {
-    it('should set isOpen to false', () => {
-      component.isOpen = true;
-
-      component.closeModal();
-
-      expect(component.isOpen).toBeFalsy();
-    });
-
-    it('should emit the close event', () => {
-      component.close = jasmine.createSpyObj({emit: null});
-
-      component.closeModal();
-
-      expect(component.close.emit).toHaveBeenCalled();
-    });
-
-  });
-});
