--- ui/src/app/modules/teams/components/modal/modal.component.ts
@@ -1,33 +1,0 @@
-/*
- * Copyright (c) 2018 Ford Motor Company
- * All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import {Component, EventEmitter, Input, Output} from '@angular/core';
-
-@Component({
-  selector: 'rq-modal',
-  templateUrl: './modal.component.html',
-  styleUrls: ['./modal.component.scss']
-})
-export class ModalComponent {
-  @Input() isOpen = false;
-  @Output() close = new EventEmitter();
-
-  closeModal() {
-    this.isOpen = false;
-    this.close.emit();
-  }
-}
