--- ui/src/app/modules/teams/components/thought-modal/thought-modal.component.html
@@ -1,37 +1,0 @@
-<!--
-  ~ Copyright (c) 2018 Ford Motor Company
-  ~ All rights reserved.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<rq-modal [isOpen]="thought != null" (close)="close()">
-  <div class="modal-header {{thought?.topic}}" id="thoughtModalHeader">
-    <div class="title">
-      {{thought?.columnTitle?.title}} thought
-    </div>
-    <div>
-      {{thought?.hearts}} <i class="fas fa-star"></i>
-    </div>
-  </div>
-
-  <div class="modal-body" id="thoughtModalBody">
-    {{thought?.message}}
-  </div>
-
-  <div class="modal-footer" id="thoughtModalFooter">
-    <input class="discussed" *ngIf="thought"
-           type="checkbox" [(ngModel)]="thought.discussed" (click)="discussThought()">
-  </div>
-</rq-modal>
-
