--- ui/src/app/modules/teams/components/thought-modal/thought-modal.component.scss
@@ -1,64 +1,0 @@
-/*!
- * Copyright (c) 2018 Ford Motor Company
- * All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-@import 'color-vars';
-
-#thoughtModalHeader {
-  &.modal-header {
-    color: $white;
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-between;
-    text-transform: uppercase;
-
-    &.happy {
-      background-color: $happy-green;
-    }
-
-    &.confused {
-      background-color: $confused-blue;
-    }
-
-    &.unhappy {
-      background-color: $unhappy-red;
-    }
-
-    .title {
-      font-size: 1.2rem;
-      max-width: calc(100% - 90px);
-    }
-  }
-}
-
-#thoughtModalBody {
-  &.modal-body {
-    font-size: 1.1rem;
-    font-weight: normal;
-  }
-}
-
-#thoughtModalFooter {
-  &.modal-footer {
-    background-color: $white;
-    padding-top: 0;
-    text-align: right;
-
-    .discussed {
-      cursor: pointer;
-      transform: scale(2);
-    }
-  }
-}
