+++ ui/src/app/modules/teams/services/column.service.ts
@@ -19,7 +19,7 @@ import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
 import {Observable} from 'rxjs/index';
 
-import {Column} from '../domain/column';
+import {Column} from '../../domain/column';
 import {WebsocketService} from './websocket.service';
 
 @Injectable()
